Fix allownil propagated to byte slices as slice elements (#374)
`[][]byte` would assume that the byte slices had allownil set. This would cause an extra `if x == nil {x = make([]byte,0)}` to be inserted on elements.

Regression from #363

+++ msgp/write_bytes_test.go
@@ -199,7 +199,7 @@ func BenchmarkAppendFloat(b *testing.B) {
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		AppendFloat64(buf, src[i&(n-1)])
+		AppendFloat(buf, src[i&(n-1)])
 	}
 }
